Move pong impl up to node
